// inventory-service/src/generated/inventory.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.6.1
//   protoc               v5.29.3
// source: inventory.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "inventory";

export interface Product {
  id: string;
  name: string;
  quantity: number;
}

export interface CheckStockRequest {
  productId: string;
  requiredQuantity: number;
}

export interface CheckStockResponse {
  isAvailable: boolean;
}

export interface UpdateStockRequest {
  productId: string;
  quantity: number;
}

export interface UpdateStockResponse {
  success: boolean;
}

export const INVENTORY_PACKAGE_NAME = "inventory";

export interface InventoryServiceClient {
  checkStock(request: CheckStockRequest): Observable<CheckStockResponse>;

  updateStock(request: UpdateStockRequest): Observable<UpdateStockResponse>;
}

export interface InventoryServiceController {
  checkStock(
    request: CheckStockRequest,
  ): Promise<CheckStockResponse> | Observable<CheckStockResponse> | CheckStockResponse;

  updateStock(
    request: UpdateStockRequest,
  ): Promise<UpdateStockResponse> | Observable<UpdateStockResponse> | UpdateStockResponse;
}

export function InventoryServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["checkStock", "updateStock"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("InventoryService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("InventoryService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const INVENTORY_SERVICE_NAME = "InventoryService";